Chapter spike (#2)
**Pull Request Description**

Creates a spike for a chapter with some stubbed in exercises to get a
feel for the structure and mechanics of creating a text.

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.