A more complete spike of key elements of the text (#3)
**Pull Request Description**

Includes some of the following elements:
- Preface link to launch the deployer
- Some examples of different question types including
  - short answer with regular expression checks
  - Parson's problems
  - multiple choice
  - matching
  - Clickable area
    - choose terms that best fit
    - Likert Scale
- README with:
  - naming convention for xml id's
  - overview of the structure of the text and sections
 

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.