Used more full featured TeX container (#5)
**Pull Request Description**

Changes to use the full TeX container in hope of fixing an issue with
missing styles when building the print versions.

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.

Signed-off-by: Grant Braught <braught@dickinson.edu>