clarifies when instructor info shows up (#11)
**Pull Request Description**

Clarifies that instructor notes appear in the `instructor` versions and
in the `web` and `print` versions. It appears in the `web` and `print`
versions so that they can be seen in the CodeChat preview.

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.