docs: update naming convention (#12)
Simplifies the naming convention for xml:id and labels, and tries to
make it so that updating a book with such labels and cross-references is
easier and less error prone.

---

**Licensing Certification**

GitKit is a [Free Cultural Work](https://freedomdefined.org/Definition)
and all accepted contributions are licensed as described in the
LICENSE.md file. This requires that the contributor holds the rights to
do so. By submitting this pull request **I certify that I satisfy the
terms of the [Developer Certificate of
Origin](https://developercertificate.org/)** for its contents.

Signed-off-by: Stoney Jackson <dr.stoney@gmail.com>